playground/faiss_faiss_tests.py

from dotenv import load_dotenv


from langchain_community.document_loaders import TextLoader
from langchain_community.vectorstores import FAISS
from langchain_openai import OpenAIEmbeddings
from langchain_text_splitters import CharacterTextSplitter
from langchain_community.document_loaders import PyPDFLoader



load_dotenv()


loader = TextLoader("./playground/state_of_the_union.txt")
documents = loader.load()
text_splitter = CharacterTextSplitter(chunk_size=1000, chunk_overlap=0)
docs = text_splitter.split_documents(documents)
embeddings = OpenAIEmbeddings()
db = FAISS.from_documents(docs, embeddings)
print(db.index.ntotal)

loader = PyPDFLoader("./playground/schankerlaubnis.pdf")
pages = loader.load_and_split()
text = pages[0].page_content
print(text)


